pyproject.toml
[project]
name = "neo4j-version"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.82.1",
    "neo4j>=6.1.0",
    "pymupdf>=1.27.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
]
